Move stuff from library header into defines file.
