+++ Pod/Core/SimpleAuth.h
@@ -6,71 +6,8 @@
 //  Copyright (c) 2013-2014 Byliner, Inc. All rights reserved.
 //
 
-extern NSString * const SimpleAuthErrorDomain;
-extern NSString * const SimpleAuthErrorStatusCodeKey;
-typedef NS_ENUM(NSUInteger, SimpleAuthError) {
-    
-    /**
-     The user cancelled authentication.
-     */
-    SimpleAuthErrorUserCancelled,
-    
-    /*
-     An error that occurred as the result of a failed network operation.
-     */
-    SimpleAuthErrorNetwork,
-    
-    /**
-     An error that originated in Accounts.framework.
-     */
-    SimpleAuthErrorAccounts,
-    
-    /**
-     Returned if SimpleAuth was able to parse response data.
-     */
-    SimpleAuthErrorInvalidData
-};
-
-/**
- Called when authorization either completes with a response or fails with an
- error. Should an error occur, response object will be nil.
- 
- @param responseObject The authorization response, or nil if an error occurred.
- @param error An error.
- 
- @see +authorize:completion:
- @see +authorize:options:completion:
- */
-typedef void (^SimpleAuthRequestHandler) (id responseObject, NSError *error);
-
-extern NSString * const SimpleAuthPresentInterfaceBlockKey;
-extern NSString * const SimpleAuthDismissInterfaceBlockKey;
-
-/**
- Called when a user interface element must be presented to continue
- authorization. This could be a UIViewController for web login, or a
- UIActionSheet for system login. All providers will have default
- implementations for the appropriate callback types. You can customize provider
- behavior by providing your own blocks.  This will be called on the main
- thread.
- 
- @see SimpleAuthPresentInterfaceBlockKey
- @see SimpleAuthDismissInterfaceBlockKey
- @see +configuration
- @see +authorize:options:completion:
- 
- @param userInterfaceElement An element that is about to be presented or
- dismissed.
- */
-typedef void (^SimpleAuthInterfaceHandler) (id userInterfaceElement);
-
-/**
- Key used to define the redirect URI for OAuth style providers.
- 
- @see +configuration
- @see +authorize:options:completion:
- */
-extern NSString * const SimpleAuthRedirectURIKey;
+@import Foundation;
+#import "SimpleAuthDefines.h"
 
 @interface SimpleAuth : NSObject
