+++ Pod/Core/SimpleAuth.m
@@ -11,14 +11,6 @@
 #import "SimpleAuthSingleSignOnProvider.h"
 #import "NSObject+SimpleAuthAdditions.h"
 
-#import <ReactiveCocoa/ReactiveCocoa.h>
-
-NSString * const SimpleAuthErrorDomain = @"SimpleAuthErrorDomain";
-NSString * const SimpleAuthPresentInterfaceBlockKey = @"present_interface_block";
-NSString * const SimpleAuthDismissInterfaceBlockKey = @"dismiss_interface_block";
-NSString * const SimpleAuthRedirectURIKey = @"redirect_uri";
-NSString * const SimpleAuthErrorStatusCodeKey = @"SimpleAuthErrorStatusCode";
-
 static SimpleAuthProvider *__currentProvider = nil;
 
 @implementation SimpleAuth
