+++ Pod/Core/SimpleAuthDefines.m
@@ -1,0 +1,15 @@
+//
+//  SimpleAuthDefines.m
+//  SimpleAuth
+//
+//  Created by Caleb Davenport on 11/30/14.
+//  Copyright (c) 2013-2015 Caleb Davenport. All rights reserved.
+//
+
+#import "SimpleAuthDefines.h"
+
+NSString *const SimpleAuthErrorDomain = @"SimpleAuthErrorDomain";
+NSString *const SimpleAuthErrorStatusCodeKey = @"SimpleAuthErrorStatusCode";
+NSString *const SimpleAuthPresentInterfaceBlockKey = @"present_interface_block";
+NSString *const SimpleAuthDismissInterfaceBlockKey = @"dismiss_interface_block";
+NSString *const SimpleAuthRedirectURIKey = @"redirect_uri";
